requirements.txt
accelerate==1.10.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
attrs==22.2.0
autopep8==2.0.4
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.0
bidict==0.23.1
binaryornot==0.4.4
bleach==6.0.0
blinker==1.8.2
briefcase==0.3.16
bs4==0.0.2
build==1.0.3
buildozer==1.5.0
cachetools==5.3.3
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cmudict==1.1.1
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.0.7
cookiecutter==2.5.0
cryptography==42.0.8
cycler==0.11.0
dataclasses-json==0.6.7
datasets==4.0.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
Distance==0.1.3
distlib==0.3.7
distro==1.9.0
Django==4.2.7
dmgbuild==1.6.1
dnspython==2.6.1
docutils==0.20.1
docx==0.2.4
ds-store==1.3.1
email_validator==2.2.0
emoji==1.2.0
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.16.3
filelock==3.13.1
fiona==1.9.6
flatbuffers==24.3.25
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
geopandas==0.14.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.31.0
googleapis-common-protos==1.63.2
googlemaps==4.10.0
grpcio==1.64.1
h11==0.14.0
hangul-jamo==1.0.1
haversine==2.8.1
hf-xet==1.1.8
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.4
image==1.5.33
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.1.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jamo==0.4.1
jedi==0.18.2
Jinja2==3.1.2
jiter==0.10.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
Kivy==2.2.1
Kivy-Garden==0.1.5
kiwisolver==1.4.4
kollocate==0.0.2
koparadigm==0.10.0
kss==6.0.5
kubernetes==30.1.0
langchain==0.2.6
langchain-chroma==0.1.2
langchain-community==0.2.6
langchain-core==0.2.11
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langchain-upstage==0.1.7
langsmith==0.1.83
llvmlite==0.40.0
lxml==5.3.0
mac-alias==2.2.2
markdown-it-py==3.0.0
markdownify==0.12.1
MarkupSafe==2.1.2
marshmallow==3.21.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclassic==0.5.4
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.9.1
notebook==6.5.3
notebook_shim==0.2.2
numba==0.57.0
numpy==1.24.1
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.107.0
openpyxl==3.1.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
oracledb==2.2.1
orjson==3.10.6
outcome==1.2.0
overrides==7.7.0
packaging==24.1
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pdf2image==1.17.0
pecab==1.0.8
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
pluggy==1.6.0
posthog==3.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.25.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.9.1
Pygments==2.14.0
pyparsing==3.0.9
pypdf==4.2.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproj==3.6.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-louvain @ git+https://github.com/taynaud/python-louvain@def91793772c3e77ab4167d175903a5365c24b4b
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
python-multipart==0.0.9
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0.2
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rsa==4.9
safetensors==0.6.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
selenium==4.13.0
Send2Trash==1.8.0
sentence-transformers==5.1.0
sh==2.0.6
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.31
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.37.2
statsmodels==0.14.2
streamlit==1.36.0
strsimpy==0.2.1
sympy==1.14.0
tenacity==8.4.2
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.2.0
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.21.4
toml==0.10.2
tomli_w==1.0.0
toolz==0.12.1
torch==2.8.0
tornado==6.2
tossi==0.3.1
tqdm==4.67.1
traitlets==5.9.0
transformers==4.55.4
trio==0.22.2
trio-websocket==0.11.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2023.3
ujson==5.10.0
Unidecode==1.4.0
uri-template==1.2.0
urllib3==2.0.3
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.24.7
watchfiles==0.22.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==12.0
Whoosh==2.7.4
widgetsnbextension==4.0.7
wrapt==1.16.0
wsproto==1.2.0
xlrd==1.2.0
xxhash==3.5.0
yarl==1.9.4
zipp==3.19.2
